% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BARLEYTerrain.R
\docType{data}
\name{BARLEYTerrain}
\alias{BARLEYTerrain}
\title{Barley terrain requirement for land evaluation}
\format{
A data frame with 6 rows and 8 columns
}
\description{
A dataset containing the terrain characteristics of the crop requirements for farming Barley.
}
\details{
The following are the factors for evaluation: 

\itemize{
\item Slope1 - Slope (\%) (1. Irrigrated agriculture, basin furrow irrigation)
\item Slope2 - Slope (\%) (2. High level of managemnet with full mechanization. )
\item Slope3 - Slope (\%) (3. Low level of managemnet animal traction or handwork.)
\item Flood - Flooding
\item Drainage4 - Drainage (Medium and fine textured soils)
\item Drainage5 - Drainage (Coarse textured soils - Sandy families)
}
}
\seealso{
\itemize{
 \item Yen, B. T., Pheng, K. S., and Hoanh, C. T. (2006). \emph{LUSET: Land Use Suitability Evaluation Tool User's Guide}. International Rice Research Institute.
 }
}
\keyword{dataset}
