% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BARLEYWaterCR.R
\docType{data}
\name{BARLEYWaterCR}
\alias{BARLEYWaterCR}
\title{Barley Water Requirement for Land Evaluation}
\format{A data frame with 3 rows and 8 columns}
\description{
A dataset containing the water characteristics of the crop requirements for farming Barley.
}
\details{
The following are the factors for evaluation:
\itemize{
 \item WgAv - Precipitation of growing cycle (mm)
 \item WmAv2	- Mean precipitation of second month (mm)
 \item WmAv3 - Mean precipitation of third month (mm)
}
}
\seealso{
\itemize{
 \item Yen, B. T., Pheng, K. S., and Hoanh, C. T. (2006). \emph{LUSET: Land Use Suitability Evaluation Tool User’s Guide}. International Rice Research Institute.
 }
}
\keyword{dataset}

