% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CABBAGETemp.R
\docType{data}
\name{CABBAGETemp}
\alias{CABBAGETemp}
\title{Cabbage temp requirement for land evaluation}
\format{
A data frame with 3 rows and 8 columns
}
\description{
A dataset containing the temp characteristics of the crop requirements for farming Cabbage.
}
\details{
The following are the factors for evaluation: 

\itemize{
\item TgAv - Mean temperature of  the growing cycle (°C)
\item TmAv0 - Mean temp. at germination (°C) (1st month)
\item TdDiff - Temp diffrence day/night ( C )
}
}
\seealso{
\itemize{
 \item Yen, B. T., Pheng, K. S., and Hoanh, C. T. (2006). \emph{LUSET: Land Use Suitability Evaluation Tool User's Guide}. International Rice Research Institute.
 }
}
\keyword{dataset}
