% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CARROTSTemp.R
\docType{data}
\name{CARROTSTemp}
\alias{CARROTSTemp}
\title{Carrots temp requirement for land evaluation}
\format{
A data frame with 2 rows and 8 columns
}
\description{
A dataset containing the temp characteristics of the crop requirements for farming Carrots.
}
\details{
The following are the factors for evaluation: 

\itemize{
\item TgAv - Mean temperature of  the growing cycle (°C)
\item TmAv0 - Mean temp. at germination (°C) (1st month)
}
}
\seealso{
\itemize{
 \item Yen, B. T., Pheng, K. S., and Hoanh, C. T. (2006). \emph{LUSET: Land Use Suitability Evaluation Tool User's Guide}. International Rice Research Institute.
 }
}
\keyword{dataset}
