% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CASSAVATemperatureCR.R
\docType{data}
\name{CASSAVATemperatureCR}
\alias{CASSAVATemperatureCR}
\title{Cassava Temperature Requirement for Land Evaluation}
\format{A data frame with 3 rows and 8 columns}
\description{
A dataset containing the temperature characteristics of the crop requirements for farming Cassava.
}
\details{
The following are the factors for evaluation:
\itemize{
 \item TyMaxAv - Mean annual maximum temperature (°C)
 \item TmMinXmAb - Absolute min temp. coldest month (°C)
 \item TgAv - Mean temperature of  the growing cycle (°C)
}
}
\seealso{
\itemize{
 \item Yen, B. T., Pheng, K. S., and Hoanh, C. T. (2006). \emph{LUSET: Land Use Suitability Evaluation Tool User’s Guide}. International Rice Research Institute.
 }
}
\keyword{dataset}

