% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CINNAMONTemp.R
\docType{data}
\name{CINNAMONTemp}
\alias{CINNAMONTemp}
\title{Cinnamon temp requirement for land evaluation}
\format{
A data frame with 1 rows and 8 columns
}
\description{
A dataset containing the temp characteristics of the crop requirements for farming Cinnamon.
}
\details{
The following are the factors for evaluation: 

\itemize{
\item TgAv - Mean temperature of  the growing cycle (°C)
}
}
\seealso{
\itemize{
 \item Yen, B. T., Pheng, K. S., and Hoanh, C. T. (2006). \emph{LUSET: Land Use Suitability Evaluation Tool User's Guide}. International Rice Research Institute.
 }
}
\keyword{dataset}
