% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COCOATemperatureCR.R
\docType{data}
\name{COCOATemperatureCR}
\alias{COCOATemperatureCR}
\title{Cocoa Temperature Requirement for Land Evaluation}
\format{A data frame with 3 rows and 8 columns}
\description{
A dataset containing the temperature characteristics of the crop requirements for farming Cocoa.
}
\details{
The following are the factors for evaluation:
\itemize{
 \item TyAv - Mean annual temperature (°C)
 \item TyMaxAv - Mean annual maximum temperature (°C)
 \item TyMaxAv - Mean annual maximum temperature (°C)
}
}
\seealso{
\itemize{
 \item Yen, B. T., Pheng, K. S., and Hoanh, C. T. (2006). \emph{LUSET: Land Use Suitability Evaluation Tool User’s Guide}. International Rice Research Institute.
 }
}
\keyword{dataset}

