% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COCONUTTemperatureCR.R
\docType{data}
\name{COCONUTTemperatureCR}
\alias{COCONUTTemperatureCR}
\title{Coconut Temperature Requirement for Land Evaluation}
\format{A data frame with 1 row and 8 columns}
\description{
A dataset containing the temperature characteristics of the crop requirements for farming Coconut.
}
\details{
The following are the factors for evaluation:
\itemize{
 \item TyAv - Mean annual temperature (°C)
}
}
\seealso{
\itemize{
 \item Yen, B. T., Pheng, K. S., and Hoanh, C. T. (2006). \emph{LUSET: Land Use Suitability Evaluation Tool User’s Guide}. International Rice Research Institute.
 }
}
\keyword{dataset}

