% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COCONUTWaterCR.R
\docType{data}
\name{COCONUTWaterCR}
\alias{COCONUTWaterCR}
\title{Coconut Water Requirement for Land Evaluation}
\format{A data frame with 3 rows and 8 columns}
\description{
A dataset containing the water characteristics of the crop requirements for farming Coconut.
}
\details{
The following are the factors for evaluation:
\itemize{
 \item WyAv - Annual precipitation (mm)
 \item WmDryLen - Length dry season (months : P < 1/2 PET)
 \item WyhAv - Mean annual rel. humidity (\%)
}
}
\seealso{
\itemize{
 \item Yen, B. T., Pheng, K. S., and Hoanh, C. T. (2006). \emph{LUSET: Land Use Suitability Evaluation Tool User’s Guide}. International Rice Research Institute.
 }
}
\keyword{dataset}

