% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COFFEEROWaterCR.R
\docType{data}
\name{COFFEEROWaterCR}
\alias{COFFEEROWaterCR}
\title{Robusta Coffee Water Requirement for Land Evaluation}
\format{A data frame with 4 rows and 8 columns}
\description{
A dataset containing the water characteristics of the crop requirements for farming Robusta Coffee.
}
\details{
The following are the factors for evaluation:
\itemize{
 \item WyAv - Annual precipitation (mm)
 \item WmDryLen - Length dry season (months : P < 1/2 PET)
 \item WmhDryXm - Mean rel. humidity dryest month (\%)
 \item WmnN5	- n/N of  the 5 dryest months
}
}
\seealso{
\itemize{
 \item Yen, B. T., Pheng, K. S., and Hoanh, C. T. (2006). \emph{LUSET: Land Use Suitability Evaluation Tool User’s Guide}. International Rice Research Institute.
 }
}
\keyword{dataset}

