% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COTTONTemp.R
\docType{data}
\name{COTTONTemp}
\alias{COTTONTemp}
\title{Cotton temp requirement for land evaluation}
\format{
A data frame with 6 rows and 8 columns
}
\description{
A dataset containing the temp characteristics of the crop requirements for farming Banana.
}
\details{
The following are the factors for evaluation: 

\itemize{
\item TgAv - Mean temperature of  the growing cycle (°C)
\item TgMaxAv - Mean max temp. of growing cycle (°C)
\item TmMaxXm - Average max. temp. warmest month (°C)
\item TmAv2 - Mean temp. crop development stage (2nd month) (°C)
\item TdAvg3 - Mean DAY temp. of flowering stage (°C)
\item TdMinN3 - Mean Night temp. of flowering stage (°C)
}
}
\seealso{
\itemize{
 \item Yen, B. T., Pheng, K. S., and Hoanh, C. T. (2006). \emph{LUSET: Land Use Suitability Evaluation Tool User's Guide}. International Rice Research Institute.
 }
}
\keyword{dataset}
