% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COTTONWater.R
\docType{data}
\name{COTTONWater}
\alias{COTTONWater}
\title{Cotton water requirement for land evaluation}
\format{
A data frame with 6 rows and 8 columns
}
\description{
A dataset containing the water characteristics of the crop requirements for farming Cotton.
}
\details{
The following are the factors for evaluation: 

\itemize{
\item WgAv - Precipitation of growing cycle (mm)
\item WmAv1 - Mean precipitation of first month (mm)
\item WmAv2 - Mean precipitation of second month (mm)
\item WmAv5 - Mean precipitation of fifth month (mm)
\item WmAv6 - Precipitation of ripening stage (mm)(6th month)
\item WmhAv3 - Relative humidity of maturation Stage (\%)
}
}
\seealso{
\itemize{
 \item Yen, B. T., Pheng, K. S., and Hoanh, C. T. (2006). \emph{LUSET: Land Use Suitability Evaluation Tool User's Guide}. International Rice Research Institute.
 }
}
\keyword{dataset}
