% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUAVATerrain.R
\docType{data}
\name{GUAVATerrain}
\alias{GUAVATerrain}
\title{Guava terrain requirement for land evaluation}
\format{
A data frame with 5 rows and 8 columns
}
\description{
A dataset containing the terrain characteristics of the crop requirements for farming Guava.
}
\details{
The following are the factors for evaluation: 

\itemize{
\item Slope1 - Slope (\%) (1. Irrigrated agriculture, basin furrow irrigation)
\item Slope2 - Slope (\%) (2. High level of managemnet with full mechanization. )
\item Slope3 - Slope (\%) (3. Low level of managemnet animal traction or handwork.)
\item Flood - Flooding
\item Drainage - Drainage
}
}
\seealso{
\itemize{
 \item Yen, B. T., Pheng, K. S., and Hoanh, C. T. (2006). \emph{LUSET: Land Use Suitability Evaluation Tool User's Guide}. International Rice Research Institute.
 }
}
\keyword{dataset}
