\docType{data}
\name{LandTerrain}
\alias{LandTerrain}
\title{Land and Terrain Chracteristics}
\format{A data frame with 2928 rows and 10 columns}
\description{
A dataset containing 2928 Land Units on 10 factors.
}
\details{
\itemize{ \item SlopeD - Slope (degree, 6 classes); \item
CFragm - Coarse fragment (Vol.%); \item SoilDpt - Soil
depth (cm); \item SoilTe - 12 classes of soil texture (Soil
Taxonomy); \item CECc - Apparent CEC Clay (cmol (+)/kg
clay); \item SumBCs - Sum of basic caions (cmol (+)/kg
soil); \item pHH20 - pH H2O; \item BS - Base Saturation
(%); \item OC - Organic carbon (%); \item Flood - Flooding;
}
}
\keyword{dataset}

