% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LaoCaiLT.R
\docType{data}
\name{LaoCaiLT}
\alias{LaoCaiLT}
\title{Land and Terrain Chracteristics of Lao Cai, Vietnam}
\format{A data frame with 2928 rows and 10 columns}
\description{
A dataset containing the land and terrain characteristics of the land units in
Lao Cai, Vietnam.
}
\details{
\itemize{
 \item SlopeD - Slope (degree, 6 classes);
 \item CFragm - Coarse fragment (Vol.\%);
 \item SoilDpt - Soil depth (cm);
 \item SoilTe - 12 classes of soil texture (Soil Taxonomy);
 \item CECc - Apparent CEC Clay (cmol (+)/kg clay);
 \item SumBCs - Sum of basic caions (cmol (+)/kg soil);
 \item pHH20 - pH H2O;
 \item BS - Base Saturation (\%);
 \item OC - Organic carbon (\%);
 \item Flood - Flooding;
}
}
\keyword{dataset}

