% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{LaoCaiTemp}
\alias{LaoCaiTemp}
\title{Temperature Chracteristics of Lao Cai, Vietnam}
\format{A data frame with 2928 rows and 12 columns}
\description{
A dataset containing the temperature characteristics of the land units
in Lao Cai, Vietnam.
}
\details{
\itemize{
 \item Jan - January;
 \item Feb - February;
 \item Mar - March;
 \item Apr - April;
 \item May - May;
 \item Jun - June;
 \item Jul - July;
 \item Aug - August;
 \item Sep - September;
 \item Oct - October;
 \item Nov - November;
 \item Dec - December;
}
}
\keyword{dataset}

