\docType{data}
\name{LUlandterrain}
\alias{LUlandterrain}
\title{landterrain}
\format{A data frame with 79 rows and 7 columns}
\source{
\href{http://countrystat.bas.gov.ph/}{Philippine Bureau of
Agricultural Statistics}
}
\description{
A dataset containing the volume of palay production in
Cordillera Administrative Region (CAR) of the Philippines,
from 1994 - 2013 quarterly.
}
\details{
\itemize{ \item Year - the time variable in quarterly.
\item Apayao - spatial locations 1.  \item Abra - spatial
location 2.  \item Kalinga - spatial location 3.  \item
Mt.Province - spatial location 4.  \item Ifugao - spatial
location 5.  \item Benguet - spatial location 6. }
}
\keyword{dataset}

