% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAIZETemperatureCR.R
\docType{data}
\name{MAIZETemperatureCR}
\alias{MAIZETemperatureCR}
\title{Maize Temperature Requirement for Land Evaluation}
\format{A data frame with 9 rows and 8 columns}
\description{
A dataset containing the Temperature characteristics of the crop requirements for farming Maize.
}
\details{
The following are the factors for evaluation:
\itemize{
 \item TgAv - Mean temperature of  the growing cycle (°C)
 \item TgMinAv - Mean min. temp. of growing cycle (°C)
}
}
\seealso{
\itemize{
 \item Yen, B. T., Pheng, K. S., and Hoanh, C. T. (2006). \emph{LUSET: Land Use Suitability Evaluation Tool User’s Guide}. International Rice Research Institute.
 }
}
\keyword{dataset}

