% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MANGOTerrain.R
\docType{data}
\name{MANGOTerrain}
\alias{MANGOTerrain}
\title{Mango terrain requirement for land evaluation}
\format{
A data frame with 6 rows and 8 columns
}
\description{
A dataset containing the terrain characteristics of the crop requirements for farming Mango.
}
\details{
The following are the factors for evaluation: 

\itemize{
\item Slope1 - Slope (\%) (1) Basin furrow irrigation
\item Slope2 - Slope (\%) (2) Mechanized, high management level
\item Slope3 - Slope (\%) (3) Manual, low management level
\item Flood - Flooding (1 - No Flood, 2 - short time, 3 - Long time)
\item Drainage 1-good, 2-moderate, 3-imperfect, 4-poor - Drainage 1-good, 2-moderate, 3-imperfect, 4-poor
\item SlopeD - 6 classes of slope (Degree)
}
}
\seealso{
\itemize{
 \item Yen, B. T., Pheng, K. S., and Hoanh, C. T. (2006). \emph{LUSET: Land Use Suitability Evaluation Tool User's Guide}. International Rice Research Institute.
 }
}
\keyword{dataset}
