% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{MarinduqueLT}
\alias{MarinduqueLT}
\title{Land and Terrain Chracteristics of Marinduque, Philippines}
\format{A data frame with 881 rows and 6 columns}
\description{
A dataset containing the land and terrain characteristics of the land units in
Marinduque, Philippines.
}
\details{
\itemize{
 \item X - Latitude of Land Units;
 \item Y - Longitude of Land Units;
 \item CECc - Apparent CEC Clay (cmol (+)/kg clay);
 \item pHH20 - pH H2O;
 \item CFragm - Coarse fragment (Vol.%);
 \item SoilTe - 12 classes of soil texture (Soil Taxonomy);
}
}
\keyword{dataset}

