% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{MarinduqueTemp}
\alias{MarinduqueTemp}
\title{Temperature Chracteristics of Marinduque, Philippines}
\format{A data frame with 881 rows and 14 columns}
\description{
A dataset containing the temperature characteristics of the land units in
Marinduque, Philippines.
}
\details{
\itemize{
 \item X - Latitude of Land Units;
 \item Y - Longitude of Land Units;
 \item Jan - January;
 \item Feb - February;
 \item Mar - March;
 \item Apr - April;
 \item May - May;
 \item Jun - June;
 \item Jul - July;
 \item Aug - August;
 \item Sep - September;
 \item Oct - October;
 \item Nov - November;
 \item Dec - December;
}
}
\keyword{dataset}

