% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarinduqueWater.R
\docType{data}
\name{MarinduqueWater}
\alias{MarinduqueWater}
\title{Water Chracteristics of Marinduque, Philippines}
\format{
A data frame with 881 rows and 14 columns
}
\description{
A dataset containing the water characteristics of the land units in
Marinduque, Philippines.
}
\details{
\itemize{
 \item Lat - Latitude of Land Units;
 \item Lon - Longitude of Land Units;
 \item Jan - January;
 \item Feb - February;
 \item Mar - March;
 \item Apr - April;
 \item May - May;
 \item Jun - June;
 \item Jul - July;
 \item Aug - August;
 \item Sep - September;
 \item Oct - October;
 \item Nov - November;
 \item Dec - December;
}
}
\keyword{dataset}
