% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MILLETSWater.R
\docType{data}
\name{MILLETSWater}
\alias{MILLETSWater}
\title{Millets water requirement for land evaluation}
\format{
A data frame with 5 rows and 8 columns
}
\description{
A dataset containing the water characteristics of the crop requirements for farming Millets.
}
\details{
The following are the factors for evaluation: 

\itemize{
\item WgAv - Precipitation of growing cycle (mm)
\item WmAv1 - Mean precipitation of first month (mm)
\item WmAv2 - Mean precipitation of second month (mm)
\item WmAv4 - Mean precipitation of fourth month (mm)
\item WmhAv4 - Relative humidity at harvest stage (\%)
}
}
\seealso{
\itemize{
 \item Yen, B. T., Pheng, K. S., and Hoanh, C. T. (2006). \emph{LUSET: Land Use Suitability Evaluation Tool User's Guide}. International Rice Research Institute.
 }
}
\keyword{dataset}
