% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OILPALMTemp.R
\docType{data}
\name{OILPALMTemp}
\alias{OILPALMTemp}
\title{Oil Palm temp requirement for land evaluation}
\format{
A data frame with 3 rows and 8 columns
}
\description{
A dataset containing the temp characteristics of the crop requirements for farming Oil Palm.
}
\details{
The following are the factors for evaluation: 

\itemize{
\item TyMaxAv - Mean annual maximum temperature (°C)
\item TdMinXm - Mean daily minimum temperature of coldest month (°C)
\item TyAv - Mean annual temperature (°C)
}
}
\seealso{
\itemize{
 \item Yen, B. T., Pheng, K. S., and Hoanh, C. T. (2006). \emph{LUSET: Land Use Suitability Evaluation Tool User's Guide}. International Rice Research Institute.
 }
}
\keyword{dataset}
