% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OILPALMTerrain.R
\docType{data}
\name{OILPALMTerrain}
\alias{OILPALMTerrain}
\title{Oil Palm terrain requirement for land evaluation}
\format{
A data frame with 4 rows and 8 columns
}
\description{
A dataset containing the terrain characteristics of the crop requirements for farming Oil Palm.
}
\details{
The following are the factors for evaluation: 

\itemize{
\item Slope - Slope
\item Flood - Flooding
\item Drainage4 - Drainage (4) Medium & fine textured soils
\item Drainage5 - Drainage (5) Coarse textured soils
}
}
\seealso{
\itemize{
 \item Yen, B. T., Pheng, K. S., and Hoanh, C. T. (2006). \emph{LUSET: Land Use Suitability Evaluation Tool User's Guide}. International Rice Research Institute.
 }
}
\keyword{dataset}
