% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OILPALMWater.R
\docType{data}
\name{OILPALMWater}
\alias{OILPALMWater}
\title{Oil Palm water requirement for land evaluation}
\format{
A data frame with 3 rows and 8 columns
}
\description{
A dataset containing the water characteristics of the crop requirements for farming Oil Palm.
}
\details{
The following are the factors for evaluation: 

\itemize{
\item WyAv - Annual precipitation (mm)
\item WmDryLen - Length dry season (months : P < 1/2 PET)
\item WynN - Mean annual n/N
}
}
\seealso{
\itemize{
 \item Yen, B. T., Pheng, K. S., and Hoanh, C. T. (2006). \emph{LUSET: Land Use Suitability Evaluation Tool User's Guide}. International Rice Research Institute.
 }
}
\keyword{dataset}
