\name{ovs}
\alias{ovs}
\title{Overall Suitability}
\usage{
ovs(dat, method = "minimum")
}
\arguments{
  \item{dat}{- the data file in data.frame class. method -
  the method to be used: minimum, maximum, exponent or
  average}
}
\description{
This functions calculates the overall suitability using the
for methods, namely minimum, maximum, exponent, and
average.
}

