% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAPAYATemp.R
\docType{data}
\name{PAPAYATemp}
\alias{PAPAYATemp}
\title{Papaya temp requirement for land evaluation}
\format{
A data frame with 2 rows and 8 columns
}
\description{
A dataset containing the temp characteristics of the crop requirements for farming Papaya.
}
\details{
The following are the factors for evaluation: 

\itemize{
\item TyAv - Mean annual temperature (°C)
\item TyMinAv - Mean annual minimum temperature (°C)
}
}
\seealso{
\itemize{
 \item Yen, B. T., Pheng, K. S., and Hoanh, C. T. (2006). \emph{LUSET: Land Use Suitability Evaluation Tool User's Guide}. International Rice Research Institute.
 }
}
\keyword{dataset}
