% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEATemp.R
\docType{data}
\name{PEATemp}
\alias{PEATemp}
\title{Pea temp requirement for land evaluation}
\format{
A data frame with 2 rows and 8 columns
}
\description{
A dataset containing the temp characteristics of the crop requirements for farming Banana.
}
\details{
The following are the factors for evaluation: 

\itemize{
\item TgAv - Mean temperature of  the growing cycle (°C)
\item TmMinAv0 - Mean min. temp. at germination (°C)
}
}
\seealso{
\itemize{
 \item Yen, B. T., Pheng, K. S., and Hoanh, C. T. (2006). \emph{LUSET: Land Use Suitability Evaluation Tool User's Guide}. International Rice Research Institute.
 }
}
\keyword{dataset}
