% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPPERGRWater.R
\docType{data}
\name{PEPPERGRWater}
\alias{PEPPERGRWater}
\title{Green Pepper water requirement for land evaluation}
\format{
A data frame with 1 rows and 8 columns
}
\description{
A dataset containing the water characteristics of the crop requirements for farming Green Pepper.
}
\details{
The following are the factors for evaluation: 

\itemize{
\item WgAv - Precipitation of growing cycle (mm)
}
}
\seealso{
\itemize{
 \item Yen, B. T., Pheng, K. S., and Hoanh, C. T. (2006). \emph{LUSET: Land Use Suitability Evaluation Tool User's Guide}. International Rice Research Institute.
 }
}
\keyword{dataset}
