% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PINEAPPLEWater.R
\docType{data}
\name{PINEAPPLEWater}
\alias{PINEAPPLEWater}
\title{Pineapple water requirement for land evaluation}
\format{
A data frame with 2 rows and 8 columns
}
\description{
A dataset containing the water characteristics of the crop requirements for farming Pineapple.
}
\details{
The following are the factors for evaluation: 

\itemize{
\item WyAv - Annual precipitation (mm)
\item WyhAv - Mean annual rel. humidity (\%)
}
}
\seealso{
\itemize{
 \item Yen, B. T., Pheng, K. S., and Hoanh, C. T. (2006). \emph{LUSET: Land Use Suitability Evaluation Tool User's Guide}. International Rice Research Institute.
 }
}
\keyword{dataset}
