% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POTATOWater.R
\docType{data}
\name{POTATOWater}
\alias{POTATOWater}
\title{Potato water requirement for land evaluation}
\format{
A data frame with 5 rows and 8 columns
}
\description{
A dataset containing the water characteristics of the crop requirements for farming Banana.
}
\details{
The following are the factors for evaluation: 

\itemize{
\item WmAv1 - Mean precipitation of first month (mm)
\item WmAv2 - Mean precipitation of second month (mm)
\item WmAv3 - Mean precipitation of third month (mm)
\item WmAv4 - Mean precipitation of fourth month (mm)
\item TgAvDlen - Average daylength growing cycle (h)
}
}
\seealso{
\itemize{
 \item Yen, B. T., Pheng, K. S., and Hoanh, C. T. (2006). \emph{LUSET: Land Use Suitability Evaluation Tool User's Guide}. International Rice Research Institute.
 }
}
\keyword{dataset}
