% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RICEBRSoilCR.R
\docType{data}
\name{RICEBRSoilCR}
\alias{RICEBRSoilCR}
\title{Rainfed Bunded Rice Soil Requirement for Land Evaluation}
\format{A data frame with 6 rows and 8 columns}
\description{
A dataset containing the Soil characteristics of the crop requirements for farming Rainfed Bunded Rice.
}
\details{
The following are the factors for evaluation:
\itemize{
 \item CFragm - Coarse fragment (Vol.\%)
 \item CECc - Apparent CEC Clay (cmol (+)/kg clay)
 \item BS - Base Saturation (\%)
 \item SumBCs - Sum of basic caions (cmol (+)/kg soil)
 \item pHH2O - pH H2O
 \item OC - Organic carbon (\%)
}
}
\seealso{
\itemize{
 \item Yen, B. T., Pheng, K. S., and Hoanh, C. T. (2006). \emph{LUSET: Land Use Suitability Evaluation Tool User’s Guide}. International Rice Research Institute.
 }
}
\keyword{dataset}

