% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RICEBRTemp.R
\docType{data}
\name{RICEBRTemp}
\alias{RICEBRTemp}
\title{RiceBR temp requirement for land evaluation}
\format{
A data frame with 4 rows and 8 columns
}
\description{
A dataset containing the temp characteristics of the crop requirements for farming RiceBR.
}
\details{
The following are the factors for evaluation: 

\itemize{
\item TgAv - Mean temperature of  the growing cycle (째C)
\item TmMaxXm - Average max. temp. warmest month (째C)
\item TmAv2 - Mean temp. crop development stage (2nd month) (째C)
\item TmMinAv4 - Average absolut Min. temperature of other months (째C)
}
}
\seealso{
\itemize{
 \item Yen, B. T., Pheng, K. S., and Hoanh, C. T. (2006). \emph{LUSET: Land Use Suitability Evaluation Tool User's Guide}. International Rice Research Institute.
 }
}
\keyword{dataset}
