% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RICEBRTerrainCR.R
\docType{data}
\name{RICEBRTerrainCR}
\alias{RICEBRTerrainCR}
\title{Rainfed Bunded Rice Terrain Requirement for Land Evaluation}
\format{A data frame with 3 rows and 8 columns}
\description{
A dataset containing the Terrain characteristics of the crop requirements for farming Rainfed Bunded Rice.
}
\details{
The following are the factors for evaluation:
\itemize{
 \item Slope1 - Slope (\%) (1. Irrigrated agriculture, basin furrow irrigation)
 \item Flood - Flooding
 \item Drainage - Drainage
}
}
\seealso{
\itemize{
 \item Yen, B. T., Pheng, K. S., and Hoanh, C. T. (2006). \emph{LUSET: Land Use Suitability Evaluation Tool User’s Guide}. International Rice Research Institute.
 }
}
\keyword{dataset}

