% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RICEIWTerrainCR.R
\docType{data}
\name{RICEIWTerrainCR}
\alias{RICEIWTerrainCR}
\title{Irrigated Rice Terrain Requirement for Land Evaluation}
\format{A data frame with 4 rows and 8 columns}
\description{
A dataset containing the Terrain characteristics of the crop requirements for farming Irrigated Rice.
}
\details{
The following are the factors for evaluation:
\itemize{
 \item Slope1 - Slope (\%) (1. Irrigrated agriculture, basin furrow irrigation)
 \item Flood - Flooding
 \item Drainage - Drainage
 \item SlopeD - Slope (degree, 6 classes)
}
}
\seealso{
\itemize{
 \item Yen, B. T., Pheng, K. S., and Hoanh, C. T. (2006). \emph{LUSET: Land Use Suitability Evaluation Tool User’s Guide}. International Rice Research Institute.
 }
}
\keyword{dataset}

