% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RICENFTemperatureCR.R
\docType{data}
\name{RICENFTemperatureCR}
\alias{RICENFTemperatureCR}
\title{Rice Cultivation Under Natural Floods Temperature Requirement for Land Evaluation}
\format{A data frame with 4 rows and 8 columns}
\description{
A dataset containing the Temperature characteristics of the crop requirements for farming Rice Cultivation Under Natural Floods.
}
\details{
The following are the factors for evaluation:
\itemize{
 \item TgAv - Mean temperature of  the growing cycle (°C)
 \item TmMaxXm - lopment stage (2nd month) (°C)
 \item TmAv2 - Mean temp. crop development stage (2nd month) (°C)
 \item TmMinAv4 - Average absolut Min. temperature of other months (°C)
}
}
\seealso{
\itemize{
 \item Yen, B. T., Pheng, K. S., and Hoanh, C. T. (2006). \emph{LUSET: Land Use Suitability Evaluation Tool User’s Guide}. International Rice Research Institute.
 }
}
\keyword{dataset}

