% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RICENFTerrain.R
\docType{data}
\name{RICENFTerrain}
\alias{RICENFTerrain}
\title{Rice Cultivation under Natural Floods terrain requirement for land evaluation}
\format{
A data frame with 2 rows and 8 columns
}
\description{
A dataset containing the terrain characteristics of the crop requirements for farming Rice Cultivation under Natural Floods.
}
\details{
The following are the factors for evaluation: 

\itemize{
\item Flood - Flooding (1 - No Flood, 2 - short time, 3 - Long time)
\item Drainage 1-good, 2-moderate, 3-imperfect, 4-poor - Drainage 1-good, 2-moderate, 3-imperfect, 4-poor
}
}
\seealso{
\itemize{
 \item Yen, B. T., Pheng, K. S., and Hoanh, C. T. (2006). \emph{LUSET: Land Use Suitability Evaluation Tool User's Guide}. International Rice Research Institute.
 }
}
\keyword{dataset}
