% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RICENFTerrainCR.R
\docType{data}
\name{RICENFTerrainCR}
\alias{RICENFTerrainCR}
\title{Rice Cultivation Under Natural Floods Terrain Requirement for Land Evaluation}
\format{A data frame with 2 rows and 8 columns}
\description{
A dataset containing the Terrain characteristics of the crop requirements for farming Rice Cultivation Under Natural Floods.
}
\details{
The following are the factors for evaluation:
\itemize{
 \item Flood	- Flooding
 \item Drainage - Drainage
}
}
\seealso{
\itemize{
 \item Yen, B. T., Pheng, K. S., and Hoanh, C. T. (2006). \emph{LUSET: Land Use Suitability Evaluation Tool User’s Guide}. International Rice Research Institute.
 }
}
\keyword{dataset}

