% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RICEURSoil.R
\docType{data}
\name{RICEURSoil}
\alias{RICEURSoil}
\title{Rainfed Upland Rice soil requirement for land evaluation}
\format{
A data frame with 8 rows and 8 columns
}
\description{
A dataset containing the soil characteristics of the crop requirements for farming Rainfed Upland Rice.
}
\details{
The following are the factors for evaluation: 

\itemize{
\item CFragm - Coarse fragment (Vol.\%)
\item SoilDpt - Soil depth (cm)
\item CECc - Apparent CEC Clay (cmol (+)/kg clay)
\item BS - Base Saturation (\%)
\item SumBCs - Sum of basic caions (cmol (+)/kg soil)
\item pHH2O - pH H2O
\item OC - Organic carbon (\%)
\item SoilTe - 12 classes of soil texture (Soil Taxonomy)
}
}
\seealso{
\itemize{
 \item Yen, B. T., Pheng, K. S., and Hoanh, C. T. (2006). \emph{LUSET: Land Use Suitability Evaluation Tool User's Guide}. International Rice Research Institute.
 }
}
\keyword{dataset}
