% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUBBERTerrain.R
\docType{data}
\name{RUBBERTerrain}
\alias{RUBBERTerrain}
\title{Rubber terrain requirement for land evaluation}
\format{
A data frame with 4 rows and 8 columns
}
\description{
A dataset containing the terrain characteristics of the crop requirements for farming Rubber.
}
\details{
The following are the factors for evaluation: 

\itemize{
\item Slope2 - Slope (\%) (2. High level of managemnet with full mechanization. )
\item Flood - Flooding
\item Drainage - Drainage
\item Slope - nan
}
}
\seealso{
\itemize{
 \item Yen, B. T., Pheng, K. S., and Hoanh, C. T. (2006). \emph{LUSET: Land Use Suitability Evaluation Tool User's Guide}. International Rice Research Institute.
 }
}
\keyword{dataset}
