% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAFFLOWERTemp.R
\docType{data}
\name{SAFFLOWERTemp}
\alias{SAFFLOWERTemp}
\title{Safflower temp requirement for land evaluation}
\format{
A data frame with 2 rows and 8 columns
}
\description{
A dataset containing the temp characteristics of the crop requirements for farming Safflower.
}
\details{
The following are the factors for evaluation: 

\itemize{
\item TmAv1 - Mean temp. of the initial stage( C )
\item TmAv2 - Mean temp. crop development stage (2nd month) (°C)
}
}
\seealso{
\itemize{
 \item Yen, B. T., Pheng, K. S., and Hoanh, C. T. (2006). \emph{LUSET: Land Use Suitability Evaluation Tool User's Guide}. International Rice Research Institute.
 }
}
\keyword{dataset}
