% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOYAWater.R
\docType{data}
\name{SOYAWater}
\alias{SOYAWater}
\title{Soya water requirement for land evaluation}
\format{
A data frame with 9 rows and 8 columns
}
\description{
A dataset containing the water characteristics of the crop requirements for farming Banana.
}
\details{
The following are the factors for evaluation: 

\itemize{
\item WgAv - Precipitation of growing cycle (mm)
\item WmAv1 - Mean precipitation of first month (mm)
\item WmAv2 - Mean precipitation of second month (mm)
\item WmAv3 - Mean precipitation of third month (mm)
\item WmAv4 - Mean precipitation of fourth month (mm)
\item WmhAv2 - Relative humidity of devel. Stage (\%)
\item WmhAv3 - Relative humidity of maturation Stage (\%)
\item WmnN2 - n/N develop. Stage (2nd month)
\item WmnN4 - n/N maturation stage (4th month)
}
}
\seealso{
\itemize{
 \item Yen, B. T., Pheng, K. S., and Hoanh, C. T. (2006). \emph{LUSET: Land Use Suitability Evaluation Tool User's Guide}. International Rice Research Institute.
 }
}
\keyword{dataset}
