% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUGARCANETemp.R
\docType{data}
\name{SUGARCANETemp}
\alias{SUGARCANETemp}
\title{SugarCane temp requirement for land evaluation}
\format{
A data frame with 4 rows and 8 columns
}
\description{
A dataset containing the temp characteristics of the crop requirements for farming SugarCane.
}
\details{
The following are the factors for evaluation: 

\itemize{
\item TdAvg0 - Mean day temperature at germination stage(°C)
\item TdAvg1 - Mean day temperature for tillage stage (°C)
\item TdAvg2 - Mean day temperature for vegetative stage (°C)
\item Tcoef - (Tmax-Tmin)/Tmean maturation stage
}
}
\seealso{
\itemize{
 \item Yen, B. T., Pheng, K. S., and Hoanh, C. T. (2006). \emph{LUSET: Land Use Suitability Evaluation Tool User's Guide}. International Rice Research Institute.
 }
}
\keyword{dataset}
