% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUGARCANEWater.R
\docType{data}
\name{SUGARCANEWater}
\alias{SUGARCANEWater}
\title{SugarCane water requirement for land evaluation}
\format{
A data frame with 4 rows and 8 columns
}
\description{
A dataset containing the water characteristics of the crop requirements for farming Banana.
}
\details{
The following are the factors for evaluation: 

\itemize{
\item Wd10 - 10 days of rainfall (mm)
\item SunH - Sunshine : hours/year
\item WynN - Mean annual n/N
\item WmhAv3 - Relative humidity of maturation Stage (\%)
}
}
\seealso{
\itemize{
 \item Yen, B. T., Pheng, K. S., and Hoanh, C. T. (2006). \emph{LUSET: Land Use Suitability Evaluation Tool User's Guide}. International Rice Research Institute.
 }
}
\keyword{dataset}
