% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEATerrain.R
\docType{data}
\name{TEATerrain}
\alias{TEATerrain}
\title{Tea terrain requirement for land evaluation}
\format{
A data frame with 4 rows and 8 columns
}
\description{
A dataset containing the terrain characteristics of the crop requirements for farming Tea.
}
\details{
The following are the factors for evaluation: 

\itemize{
\item Slope2 - Slope (\%) (2. High level of managemnet with full mechanization. )
\item Flood - Flooding (1 - No Flood, 2 - short time, 3 - Long time)
\item Drainage 1-good, 2-moderate, 3-imperfect, 4-poor - Drainage 1-good, 2-moderate, 3-imperfect, 4-poor
\item SlopeD - Slope (1: 0-3, 2: 3-8, 3: 8-15, 4: 15-20, 5: 20-25, 6: >25) (degree, 6 classes)
}
}
\seealso{
\itemize{
 \item Yen, B. T., Pheng, K. S., and Hoanh, C. T. (2006). \emph{LUSET: Land Use Suitability Evaluation Tool User's Guide}. International Rice Research Institute.
 }
}
\keyword{dataset}
