\docType{data}
\name{Temperature}
\alias{Temperature}
\title{Monthly average temperature , in Celsius}
\format{A data frame with 2928 rows and 12 columns}
\description{
A dataset containing 2928 Land Units on 12 factors.
}
\details{
\itemize{ \item Jan -January; \item Feb - February; \item
Mar - March; \item Apr - April; \item May - May; \item Jun
- June; \item Jul - July; \item Aug - August; \item Sep -
September; \item Oct - October; \item Nov - November; \item
Dec - December; }
}
\keyword{dataset}

