% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOBACCOWater.R
\docType{data}
\name{TOBACCOWater}
\alias{TOBACCOWater}
\title{Tobacco water requirement for land evaluation}
\format{
A data frame with 3 rows and 8 columns
}
\description{
A dataset containing the water characteristics of the crop requirements for farming Tobacco.
}
\details{
The following are the factors for evaluation: 

\itemize{
\item WgAv - Precipitation of growing cycle (mm)
\item WynN - Mean annual n/N
\item WghAv - Relative humidity growing cycle (\%)
}
}
\seealso{
\itemize{
 \item Yen, B. T., Pheng, K. S., and Hoanh, C. T. (2006). \emph{LUSET: Land Use Suitability Evaluation Tool User's Guide}. International Rice Research Institute.
 }
}
\keyword{dataset}
