% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOMATOTemp.R
\docType{data}
\name{TOMATOTemp}
\alias{TOMATOTemp}
\title{Tomato temp requirement for land evaluation}
\format{
A data frame with 4 rows and 8 columns
}
\description{
A dataset containing the temp characteristics of the crop requirements for farming Tomato.
}
\details{
The following are the factors for evaluation: 

\itemize{
\item TgAv - Mean temperature of  the growing cycle (째C)
\item TmAv0 - Mean temp. at germination (째C) (1st month)
\item TmAv3 - Mean temp. of the flowering stage (째C)
\item TdDiff3 - Temp. diff day/night flowering stage (째C)
}
}
\seealso{
\itemize{
 \item Yen, B. T., Pheng, K. S., and Hoanh, C. T. (2006). \emph{LUSET: Land Use Suitability Evaluation Tool User's Guide}. International Rice Research Institute.
 }
}
\keyword{dataset}
