\docType{data}
\name{Water}
\alias{Water}
\title{Monthly average rainfall, mm and irrigation status}
\format{A data frame with 2928 rows and 13 columns}
\description{
A dataset containing 2928 Land Units on 13 factors.
}
\details{
\itemize{ \item Jan -January; \item Feb - February; \item
Mar - March; \item Apr - April; \item May - May; \item Jun
- June; \item Jul - July; \item Aug - August; \item Sep -
September; \item Oct - October; \item Nov - November; \item
Dec - December; \item Irrigation - Irrigation; }
}
\keyword{dataset}

