% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WHEATTemp.R
\docType{data}
\name{WHEATTemp}
\alias{WHEATTemp}
\title{Wheat temp requirement for land evaluation}
\format{
A data frame with 4 rows and 8 columns
}
\description{
A dataset containing the temp characteristics of the crop requirements for farming Wheat.
}
\details{
The following are the factors for evaluation: 

\itemize{
\item TgAv - Mean temperature of  the growing cycle (째C)
\item TmAv2 - Mean temp. crop development stage (2nd month) (째C)
\item TmAv3 - Mean temp. of the flowering stage (째C)
\item TmAv4 - Mean temp. of the ripening stage (째C)
}
}
\seealso{
\itemize{
 \item Yen, B. T., Pheng, K. S., and Hoanh, C. T. (2006). \emph{LUSET: Land Use Suitability Evaluation Tool User's Guide}. International Rice Research Institute.
 }
}
\keyword{dataset}
