% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WHEATWater.R
\docType{data}
\name{WHEATWater}
\alias{WHEATWater}
\title{Wheat water requirement for land evaluation}
\format{
A data frame with 4 rows and 8 columns
}
\description{
A dataset containing the water characteristics of the crop requirements for farming Banana.
}
\details{
The following are the factors for evaluation: 

\itemize{
\item WgAv - Precipitation of growing cycle (mm)
\item WmAv2 - Mean precipitation of second month (mm)
\item WmAv3 - Mean precipitation of third month (mm)
\item WmAv4 - Mean precipitation of fourth month (mm)
}
}
\seealso{
\itemize{
 \item Yen, B. T., Pheng, K. S., and Hoanh, C. T. (2006). \emph{LUSET: Land Use Suitability Evaluation Tool User's Guide}. International Rice Research Institute.
 }
}
\keyword{dataset}
