% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/LaoCaiWater.R
\docType{data}
\name{LaoCaiWater}
\alias{LaoCaiWater}
\title{Water Chracteristics of Lao Cai, Vietnam}
\format{A data frame with 2928 rows and 12 columns}
\description{
A dataset containing the water chracteristics of the land units in Lao Cai, Vietnam.
}
\details{
\itemize{
 \item Jan - January;
 \item Feb - February;
 \item Mar - March;
 \item Apr - April;
 \item May - May;
 \item Jun - June;
 \item Jul - July;
 \item Aug - August;
 \item Sep - September;
 \item Oct - October;
 \item Nov - November;
 \item Dec - December;
}
}
\keyword{dataset}

