% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/map_suit.R
\name{map_suit}
\alias{map_suit}
\title{Mapping Suitability Scores and Class}
\usage{
map_suit(x, y, location, suit = NULL, ovsuit = FALSE, method = NULL,
  interval = NULL, char = NULL, cols = 2, size = 4, alpha = 1,
  extent = "panel", base_layer, maprange = FALSE, padding = 0.02,
  darken = c(0, "black"), ...)
}
\arguments{
\item{x}{an object of class suitability;}

\item{y}{a data frame consisting the properties of the land units;}

\item{location}{location of the land units, an address, longitude/latitude pair (in that order),
argument of the \code{\link{get_map}} function.}

\item{suit}{suitability type to be overlayed on the map, either \code{"scores"} or
\code{"class"} of the land units. If \code{NULL}, scores of the suitability are mapped.}

\item{ovsuit}{if \code{TRUE}, the overall suitability of the land units are included in the map.}

\item{method,interval}{arguments of the \code{\link{overall_suit}}.}

\item{char}{characteristic(s)/factor(s) of the land units to be mapped. If \code{NULL},
all factors are plotted. So for the first factor of the land units
assign \code{char = 1}, for first two factors of the land units, assign
\code{char = 1:2} or \code{char = c(1, 2)}, and so on. Note that, the number of
factors depends on the number of characteristics in the crop requirements that
matches with the characteristics of the land units.}

\item{cols}{number of columns in the layout of the map.}

\item{size}{size of the points overlayed in the map.}

\item{alpha}{transparency of the points overlayed in the map.}

\item{extent}{argument of the \code{\link{get_map}} function.}

\item{base_layer,}{}

\item{maprange,}{arguments of the \code{\link{ggmap}}
function.}

\item{padding,darken}{}

\item{...}{further argument of the \code{\link{get_map}} function.}
}
\description{
This function maps the scores and class of the suitability for all land units.
}
\details{
The function depends on the ggmap package, thus themes of the map can be configured
using the \code{\link{theme}} function of the ggplot package.
}
\examples{
library(ALUES)
x <- MarinduqueLT
y <- COCONUTSoilCR
coconut_tersuit <- suitability(x = x, y = y, interval = "unbias")

# Map suitability class
map_suit(x = coconut_tersuit, y = x, suit = "class",
  location = c(121.999659, 13.378625), zoom = 11)

x <- MarinduqueWater
y <- RICEBRWaterCR
rice_watsuit <- suitability(x = x, y = y,
  interval = "unbias", sow.month = 1)

# Map suitability scores
map_suit(x = rice_watsuit, y = x, suit = "scores",
  location = c(121.999659, 13.378625), zoom = 11)

# Map suitability class
map_suit(x = rice_watsuit, y = x, suit = "class",
  location = c(121.999659, 13.378625), zoom = 11)

# Map suitability scores with overall suitability
map_suit(x = rice_watsuit, y = x, suit = "scores",
  ovsuit = TRUE, location = c(121.999659, 13.378625), zoom = 11)
}
\seealso{
\code{\link{suitability}}, \code{\link{overall_suit}}.
}

