% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overall_suit.R
\name{overall_suit}
\alias{overall_suit}
\title{Overall Suitability Scores/Class of the Land Units}
\usage{
overall_suit(suit, method = NULL, interval = NULL)
}
\arguments{
\item{suit}{an object of class suitability.}

\item{method}{the method for computing the overall suitability, choices are:
\code{"minimum"}, \code{"maximum"}, and
\code{"average"}. If \code{NULL}, method is set to \code{"minimum"}.}

\item{interval}{if \code{NULL}, the interval of the suitability class are the following: 0\% - 25\% (Not
suitable, N), 25\% - 50\% (Marginally Suitable, S3), 50\% - 75\% (Moderately Suitable, S2), and
75\% - 100\% (Highly Suitable, S1). But users can assign custom intervals by specifying
the values of the end points of the intervals. Say for intervals: 0\% - 20\% (Not
suitable, N), 20\% - 50\% (Marginally Suitable, S3), 50\% - 80\% (Moderately Suitable, S2), and
80\% - 100\% (Highly Suitable, S1), is equivalent to \code{interval = c(0, 0.2, 0.5, 0.8, 1)}.}
}
\description{
This function computes the overall suitability scores and class of the land units.
}
\examples{
library(ALUES)
x <- LaoCaiLT
y <- COCONUTSoil

# Compute the suitability of the land units of the Lao Cai land
# terrain characteristics using the 
coconut_tersuit <- suitability(x = x, y = y)

# Return the first 10 of the observations
lapply(coconut_tersuit, function(x) head(x, n = 10))

# Compute the overall suitability of the characteristics
head(overall_suit(coconut_tersuit))
}
\seealso{
\code{\link{suitability}}
}
