% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/overall_suit.R
\name{overall_suit}
\alias{overall_suit}
\title{Overall Suitability Scores/Class of the Land Units}
\usage{
overall_suit(x, method = c("minimum", "maximum", "sum", "product", "mean",
  "exponential"), interval = NULL, output = NULL)
}
\arguments{
\item{x}{a data frame consisting the suitability scores of a given characteristics
(terrain, soil, water and temperature) for a
given crop (e.g. coconut, cassava, etc.);}

\item{method}{the method for computing the overall suitability, which includes the
\code{"minimum"}, \code{"maximum"}, \code{"sum"}, \code{"product"}, and
\code{"average"}.}

\item{interval}{if \code{NULL}, the interval used are the following: 0-25% (Not
suitable, N), 25%-50% (Marginally Suitable, S3), 50%-75% (Moderately Suitable, S2), and
75%-100% (Highly Suitable, S1). But users can assign a custom intervals by specificying
the values of the end points of the intervals.}

\item{output}{the output to be returned, either the scores or class. If \code{NULL},
       both are returned.}
}
\description{
This function computes the overall suitability scores and class of the land units.
}
\examples{
library(ALUES)
x <- LaoCaiLT
y <- COCONUTSoilCR

coconut_tersuit <- suitability(x = x, y = y)
lapply(coconut_tersuit, function(x) head(x, n = 10))

head(overall_suit(coconut_tersuit[[2]]))
}

