% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suitability.R
\name{suitability}
\alias{suitability}
\title{Suitability Scores/Class of the Land Units}
\usage{
suitability(
  x,
  y,
  mf = "triangular",
  sow.month = NULL,
  min = NULL,
  max = "average",
  interval = NULL,
  sigma = NULL
)
}
\arguments{
\item{x}{a data frame consisting the properties of the land units;}

\item{y}{a data frame consisting the requirements of a given 
characteristics (terrain, soil, water and temperature) for a 
given crop (e.g. coconut, cassava, etc.);}

\item{mf}{membership function with default assigned to \code{"triangular"} 
fuzzy model. Other fuzzy models included are \code{"trapezoidal"} and
\code{"gaussian"}.}

\item{sow.month}{sowing month of the crop. Takes integers from 1 to 12 
(inclusive), representing the twelve months of a year. 
So if sets to 1, the function assumes sowing month on 
January.}

\item{min}{factor's minimum value. If \code{NULL} (default), \code{min} is
set to 0. But if numeric of length one, say 0.5, then minimum 
is set to 0.5, for all factors. If factors on land units 
(\code{x}) have different minimum, then these can be concatenated
to vector of \code{min}s, the length of this vector should be equal
to the number of factors in \code{x}. However, if set to \code{"average"},
then \code{min} is computed from different conditions:

If for example, using \code{ALUES::COCONUTSoil} (coconut terrain requirements), 
as shown below,

\code{     code s3_a s2_a  s1_a s1_b s2_b s3_b          wts}\cr
\code{1  CFragm 55.0 35.0  15.0   NA   NA   NA           NA}\cr
\code{2 SoilDpt 50.0 75.0 100.0   NA   NA   NA           NA}\cr
\code{3      BS 19.9 19.9  20.0   NA   NA   NA           NA}\cr
\code{4  SumBCs  1.5  1.5   1.6   NA   NA   NA           NA}\cr
\code{5      OC  0.7  0.7   0.8   NA   NA   NA           NA}\cr
\code{6   ECemh 20.0 16.0  12.0   NA   NA   NA           NA}}

\item{max}{maximum value for factors. Default is to \code{"average"}, check on
the details for this option. Assignment on maximum can also be done by
simply entering any real numbers, say 55, then max is 55, we say this is homogeneous,
since the maximum value for all factors then is set to 55. But for heterogeneous. For  \code{max}
on every factor, simply concatenate the different \code{max} for each factor. 
If set to \code{"average"}, check on details below for more.}

\item{interval}{domains for every suitability class (S1, S2, S3). If fixed, the
interval would be 0 to 25\% for N (Not Suitable), 25\% to 50\% for S3 (Marginally Suitable),
50\% to 75\% for S2 (Moderately Suitable), and 75\% to 100\% for (Highly Suitable).}

\item{sigma}{If \code{mf = "gaussian"}, then sigma represents the constant sigma in the
gaussian formula, which is often times referred as the variance.}
}
\description{
This function calculates the suitability scores and class of the land units.
}
\details{
There are four membership functions and these are triangular, trapezoidal, gaussian, and sigmoidal.
For triangular case. If a given factor has values equal for all suitabilities, then the class will 
trimmed down to N (not suitable) with domain [0, max), and S1 (highly suitable) with single tone domain \{0\}.
}
\examples{
library(ALUES)
x <- LaoCaiLT
y <- COCONUTSoil

coconut_tersuit <- suitability(x = x, y = y)
lapply(coconut_tersuit, function(x) head(x, n = 10))
}
